requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.
pbr>=2.0 # Apache-2.0
PyYAML>=3.10.0 # MIT
six>=1.9.0 # MIT
voluptuous>=0.8.9 # BSD License
ansible>=2.3.0.0 # GPLv3
